+++ packages/yourloops/components/team/team-monitoring-alerts-configuration.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2025, Diabeloop
+ * Copyright (c) 2022-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -45,6 +45,7 @@ import {
 import { useParams } from 'react-router-dom'
 import { logError } from '../../utils/error.util'
 import { errorTextFromException } from '../../lib/utils'
+import CardHeader from '@mui/material/CardHeader'
 
 export interface TeamMonitoringAlertsConfigurationProps {
   team: Team
@@ -85,12 +86,10 @@ function TeamMonitoringAlertsConfiguration(props: TeamMonitoringAlertsConfigurat
   return (
     <div className={commonTeamClasses.root} data-testid="team-monitoring-alerts-configuration">
       <div className={commonTeamClasses.categoryHeader}>
-        <div data-stonlyid="care-team-settings-monitoring-alerts-configuration-title">
-          <TuneIcon />
-          <Typography className={commonTeamClasses.title} data-testid="monitoring-alert-configuration-section">
-            {t('monitoring-alerts-configuration')}
-          </Typography>
-        </div>
+        <CardHeader
+          title={t('monitoring-alerts-configuration')}
+          data-testid="monitoring-alert-configuration-section"
+        />
       </div>
       <Box sx={{ paddingX: 3 }}>
         <MonitoringAlertsTeamConfiguration